+++ backend/src/main/java/com/shinhan/walfi/service/CharacterServiceImpl.java
@@ -1,0 +1,96 @@
+package com.shinhan.walfi.service;
+
+import com.shinhan.walfi.domain.CharacterType;
+import com.shinhan.walfi.domain.game.GameCharacter;
+import com.shinhan.walfi.domain.game.UserGameInfo;
+import com.shinhan.walfi.repository.CharacterRepository;
+import com.shinhan.walfi.repository.UserGameInfoRepository;
+import com.shinhan.walfi.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Random;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class CharacterServiceImpl implements CharacterService {
+
+    private final CharacterRepository characterRepository;
+    private final UserRepository userRepository;
+    private final UserGameInfoRepository userGameInfoRepository;
+
+    /**
+     * 캐릭터 타입 랜덤 설정 후 캐릭터 생성
+     *
+     * @param userId
+     * @return 생성한 캐릭터 idx 반환
+     */
+    @Override
+    @Transactional
+    public Long create(String userId) {
+        UserGameInfo userGameInfo = userRepository.findUserGameInfo(userId);
+
+        Random random = new Random();
+
+        // 캐릭터 타입 랜덤 생성
+        CharacterType[] characterTypes = CharacterType.values();
+        int typesRandomNum = random.nextInt(characterTypes.length);
+        CharacterType randomCharacterType = characterTypes[typesRandomNum];
+
+        // 캐릭터 생성
+        Boolean isMain = true;
+        GameCharacter gameCharacter = GameCharacter.createCharacter(userGameInfo, randomCharacterType, isMain);
+
+        // db에 저장
+        characterRepository.save(gameCharacter);
+
+        return gameCharacter.getCharacterIdx();
+    }
+
+    /**
+     * 캐릭터 뽑기
+     *
+     * @param userId
+     * @return 뽑은 캐릭터 idx 반환
+     */
+    @Override
+    @Transactional
+    public Long shop(String userId) {
+        UserGameInfo userGameInfo = userRepository.findUserGameInfo(userId);
+
+        Random random = new Random();
+
+        // 캐릭터 타입 랜덤 생성
+        CharacterType[] characterTypes = CharacterType.values();
+        int typesRandomNum = random.nextInt(characterTypes.length);
+        CharacterType randomCharacterType = characterTypes[typesRandomNum];
+
+        // 캐릭터 생성
+        Boolean isMain = false;
+        GameCharacter gameCharacter = GameCharacter.createCharacter(userGameInfo, randomCharacterType, isMain);
+
+        // db에 저장
+        characterRepository.save(gameCharacter);
+
+        return gameCharacter.getCharacterIdx();
+    }
+
+    /**
+     * 사용자가 가진 전체 캐릭터 조회
+     *
+     * @param userId
+     * @return 캐릭터 DTO
+     */
+    @Override
+    public List<GameCharacter> searchCharacters(String userId) {
+        UserGameInfo userGameInfo = userGameInfoRepository.findByUserId(userId);
+
+        List<GameCharacter> characters = characterRepository.findCharactersByUserGameInfo(userGameInfo);
+
+        return characters;
+    }
+
+}
